Update the docs; rename Api.js => ApiClient.js
